Добавляет новый модуль для отрисовки миниатюр и переносит весь код из модуля image-viewer.js в него
